Change to activity Moved to view
