+++ app/src/main/java/com/metrostateics499/vre_app/view/VerificationCodeActivity.kt
@@ -1,4 +1,4 @@
-package com.metrostateics499.vre_app
+package com.metrostateics499.vre_app.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,8 +7,9 @@ import android.widget.EditText
 import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.metrostateics499.vre_app.R
 
-class VerificationCode : AppCompatActivity() {
+class VerificationCodeActivity : AppCompatActivity() {
 
     private lateinit var enterCode: EditText
     private lateinit var textEmailUsernameHint: TextView
@@ -53,7 +54,7 @@ class VerificationCode : AppCompatActivity() {
             Toast.makeText(this, "Continue to Set New Password", Toast.LENGTH_LONG).show()
 
             val intent =
-                Intent(this, SetNewPassword::class.java)
+                Intent(this, SetNewPasswordActivity::class.java)
                     .putExtra("email/username", username)
             startActivity(intent)
         } else {
