+++ app/src/main/res/layout/activity_set_new_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FF185CA5"
-    tools:context=".SetNewPassword">
+    tools:context=".view.SetNewPasswordActivity">
 
     <TextView
         android:id="@+id/app_logo"
