+++ app/src/main/res/layout/activity_verification_code.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FF185CA5"
-    tools:context=".VerificationCode">
+    tools:context=".view.VerificationCodeActivity">
 
     <TextView
         android:id="@+id/app_logo"
@@ -112,7 +112,7 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@id/enter_code" />
-    
+
 
     <ImageView
         android:id="@+id/maverick_image"
